refactor(contracts): split types, rename ContributorGate, migrate registry
- Split crowdfund/src/types.rs (1 756 lines, 126 types) into
  types/domain.rs (storage/state types, DataKey) and
  types/events.rs (all 35+ Event* structs), with types/mod.rs
  re-exporting both. No public type signatures changed.

- Rename crowdfund::security::AccessControl -> ContributorGate to
  eliminate the name collision with common::AccessControl
  (the role-auth primitive). Update re-export in lib.rs and
  fix tests in security.rs.

- Add common/src/validation.rs::validate_positive_amount as the
  single canonical 'amount > 0' check. Both callers now delegate:
    achievements/src/validation.rs::validate_amount
    crowdfund/src/validation.rs::validate_positive_amount

- Finish registry migration to contracts/common per ADR-004:
    registry/src/errors.rs: add From<CommonError> for ContractError
    registry/src/admin.rs: use AccessControl::require_role_auth in
    update_status (replaces manual admin.require_auth() inline pattern)

- Update ADR-004 status Proposed -> Active; confirm dead rbac*.rs
  blocker is gone (da3d8d0) and all acceptance criteria are met.
