+++ contracts/common/src/lib.rs
@@ -9,6 +9,7 @@
 mod access_control;
 mod error;
 mod events;
+mod validation;
 
 #[cfg(any(test, feature = "testutils"))]
 pub mod test_utils;
@@ -16,3 +17,4 @@ pub mod test_utils;
 pub use access_control::AccessControl;
 pub use error::CommonError;
 pub use events::EVENT_SCHEMA_VERSION;
+pub use validation::validate_positive_amount;
